0.3

+++ options.php
@@ -5,10 +5,12 @@
 
     <?php  
         $active_tab = isset( $_GET[ 'tab' ] ) ? $_GET[ 'tab' ] : 'welcome';  
+        $root = dirname(__FILE__);
     ?>  
 
     <h2 class="nav-tab-wrapper">  
-        <a href="?page=sslcom_cc&tab=welcome" class="nav-tab <?php echo $active_tab == 'welcome' ? 'nav-tab-active' : ''; ?>">Welcome</a>  
+        <a href="?page=sslcom_cc&tab=welcome" class="nav-tab <?php echo $active_tab == 'welcome' ? 'nav-tab-active' : ''; ?>">Welcome</a>
+        <a href="?page=sslcom_cc&tab=bundle" class="nav-tab <?php echo $active_tab == 'bundle' ? 'nav-tab-active' : ''; ?>">Bundles</a>  
         <a href="?page=sslcom_cc&tab=html" class="nav-tab <?php echo $active_tab == 'html' ? 'nav-tab-active' : ''; ?>">HTML</a>  
         <a href="?page=sslcom_cc&tab=php" class="nav-tab <?php echo $active_tab == 'php' ? 'nav-tab-active' : ''; ?>">PHP</a>
         <a href="?page=sslcom_cc&tab=css" class="nav-tab <?php echo $active_tab == 'css' ? 'nav-tab-active' : ''; ?>">CSS</a>
@@ -16,17 +18,29 @@
     </h2>
 
     <?php
-        if (isset($_GET['edit'])) include_once(dirname(__FILE__) . "/fragments/edit_code.php");
-        elseif (isset($_GET['delete']) && $_GET['delete'] !== '') {
-            $file = dirname(__FILE__) . "/".$active_tab."/".$_GET['delete'];
+        if (isset($_GET['edit'])) include_once(correctpath($root . "/fragments/edit_code.php"));
+        elseif (isset($_GET['delete']) && $_GET['delete'] !== '' && $active_tab !== 'bundle') {
+            $file = correctpath($root . "/" . $active_tab . "/" . $_GET['delete']);
             if (file_exists($file)) {
                 unlink($file);
-                include_once(dirname(__FILE__) . '/' . "list_code.php");
+                echo '<div class="updated"><p>$file has been deleted</p></div>';
+                include_once(correctpath($root . '/fragments/list_code.php'));
             }
         }
         else {
-        	if($active_tab !== "welcome") include_once(dirname(__FILE__) . '/' . "list_code.php");
-        	else include_once(dirname(__FILE__) . "/fragments/welcome.php");
+            switch ($active_tab) {
+                case "welcome" :
+                    include_once(correctpath($root . "/fragments/welcome.php"));
+                    break;
+
+                case "bundle" :
+                    include_once(correctpath($root . "/fragments/bundle_handler.php"));
+                    break;
+
+                default :
+                    include_once(correctpath($root . '/fragments/list_code.php'));
+                    break;
+            }
         }
 	?>
 </div>
